+++ WinAPI/Calc/Calc.vcxproj
@@ -130,7 +130,7 @@
     <ClCompile Include="main.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="Colors.h" />
+    <ClInclude Include="Skins.h" />
     <ClInclude Include="Dimensions.h" />
     <ClInclude Include="Fonts.h" />
     <ClInclude Include="resource.h" />
