+++ WinAPI/Calc/Calc.vcxproj.filters
@@ -26,7 +26,7 @@
     <ClInclude Include="Dimensions.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="Colors.h">
+    <ClInclude Include="Skins.h">
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="Fonts.h">
