+++ WinAPI/Calc/Dimensions.h
@@ -6,7 +6,7 @@ CONST INT g_i_INTERVAL = 5;		// расстояние между кнопками
 CONST INT g_i_BUTTON_DOUBLE_SIZE = g_i_BUTTON_SIZE * 2 + g_i_INTERVAL; // размер кнопки в пикселах
 
 CONST INT g_i_SCREEN_WIDTH = g_i_BUTTON_SIZE * 5 + g_i_INTERVAL * 4;
-CONST INT g_i_SCREEN_HEIGHT = 44;
+CONST INT g_i_SCREEN_HEIGHT = 48;
 CONST INT g_i_FONT_HEIGHT = g_i_SCREEN_HEIGHT *0.8;
 CONST INT g_i_FONT_WIDTH = g_i_SCREEN_HEIGHT / 2;
