+++ WinAPI/Calc/Fonts.h
@@ -3,7 +3,7 @@
 
 CONST CHAR* g_FONT_NAMES[] =
 {
-	"Digital - 7",
+	"Digital-7",
    "MOSCOW2024",
-	"terminat"
+	"Terminator Two"
 };
