+++ WinAPI/Calc/Skins.h
